# Restructure `sum` for better auto-vectorization