@jhorstmann 2023-07-22
Interesting, I can reproduce the benchmark results on my machine (i9-11900KB with AVX-512), the version without simd feature is even slightly faster. Very nice improvement!

With `target-cpu=skylake` there is still a small difference for the nullable version, 70ns vs 56ns with simd feature. The packed_simd code is not taking full advantage of avx512 mask registers and therefore runs at the same speed when targeting either skylake or native.

What cpu did you run your benchmarks on?