@simonvandel 2023-07-22
> Interesting, I can reproduce the benchmark results on my machine (i9-11900KB with AVX-512), the version without simd feature is even slightly faster. Very nice improvement!

Great! Do you mean that both the non-null and null version are competitive with the simd feature on your machine?

> What cpu did you run your benchmarks on?

It's an i7-10750H. I used the following rustc: `rustc 1.73.0-nightly (0308df23e 2023-07-21)`