@jhorstmann 2023-07-23
These are my results, simd feature is a tiny bit ahead, but maybe not enough to justify the additional code complexity.

```
$ RUSTFLAGS='-C target-cpu=native' cargo +nightly bench --bench aggregate_kernels "sum"
sum 512                 time:   [20.207 ns 20.244 ns 20.285 ns]
sum nulls 512           time:   [58.970 ns 59.000 ns 59.035 ns]

$ RUSTFLAGS='-C target-cpu=native' cargo +nightly bench --features simd --bench aggregate_kernels "sum"
sum 512                 time:   [17.095 ns 17.107 ns 17.120 ns]
sum nulls 512           time:   [56.853 ns 56.887 ns 56.925 ns]
```