@simonvandel 2023-07-23
I'll let you/others decide if we should replace the simd feature impl with the code in this PR. And if so, should this be done in this PR, or another one?

I'll push a commit today it tomorrow that resolves the todo, picking a proper value for LANES based on T.