@simonvandel 2023-07-30
I couldn't find a single implementation that would speed up both integer and floating point numbers, so I decided to have an implementation for both.
This should keep the current performance for all integer types, but give significant speed-ups for floating points.

Marked ready to review.